Finally done with Symbolics.
At least almost one, maybe we have to revisit this when we want to
evaluate this, idk if this path is completely bug free.

But generally we ended up with a pretty powerful solver, which
allows proving statements for us like the
equality of two tensor extents, or proving the alignment of tensors.

Just to give an example it's able to prove that given a aligned variable
E if we pool and follow the pool with a upsample step, we actually end
up with the original variable which is exactly what we would expect, if
E is aligned by at least 2. This even works for multiple pool steps,
followed by upsampling.

Missing features, there are a couple of uncommon cases where the
denominator is a symbolic or a affine expression of symbolics where we
cannot prove some statements, but iam actually totally fine with that.

We also do not yet have any sense or ordering, which would maybe be a
reasonable thing to integrate at some point.
Similarly the operations min / max would be nice, but they would of
cause require those constraits.
Maybe a simple version could just look at two affine expressions and
then be able to reduce operations like min / max into noops, but not
sure. I think this will maybe be useful if we ever want to do proper
memory planning while considering lifetimes,
but this is something that we can do after we have a first working
prototype.